Adding helper function to get latest INFO log
